+++ CHANGELOG.md
@@ -10,6 +10,14 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 ## Note - after update to 4.7.x buttons, sensors, volume display type need to be configure again using config UI
 ## Note - after update to 3.15.x need remove the accessory frome Home app and add it again
 
+## [5.0.1] - (14.08.2024)
+
+## Changes
+
+- performance improvements and stability
+- fix mqtt and restFul start in some cases
+- cleanup
+
 ## [5.0.0] - (14.08.2024)
 
 ## Changes